remove unnecessary casting

+++ matchbox_demo/src/box_game.rs
@@ -155,7 +155,7 @@ pub fn move_cube_system(
     inputs: Res<PlayerInputs<GGRSConfig>>,
 ) {
     for (mut t, mut v, p) in query.iter_mut() {
-        let input = inputs[p.handle as usize].0.inp;
+        let input = inputs[p.handle].0.inp;
         // set velocity through key presses
         if input & INPUT_UP != 0 && input & INPUT_DOWN == 0 {
             v.z -= MOVEMENT_SPEED;
